backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
pytest
pytest-asyncio
bcrypt
pyjwt
psycopg2-binary
sqlalchemy
pydantic[email]
email-validator
